requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.7
celery==5.4.0
prometheus-client==0.20.0
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
black==24.8.0
isort==5.13.2
